+++ lib/oli/delivery/gating/condition_types/condition_type.ex
@@ -3,6 +3,7 @@ defmodule Oli.Delivery.Gating.ConditionTypes.ConditionType do
   Behavior a gating condition type must implement
   """
   alias Oli.Delivery.Gating.GatingCondition
+  alias Oli.Delivery.Gating.ConditionTypes.ConditionContext
 
   @doc """
   Returns the condition type identifier
@@ -10,12 +11,12 @@ defmodule Oli.Delivery.Gating.ConditionTypes.ConditionType do
   @callback type() :: Atom.t()
 
   @doc """
-  Returns true if the condition evaluation determines that the resource can be accessed.
-  This function call is inteded to be as efficient as possible for the common case when
-  a resource is accessible. Use the `details` callback to get more details as to
-  why a resource is inaccessible.
+  Returns a two element tuple. The first element is a boolean indicating whether this resource
+  can be accessed strictly based on this gating condition.  The second element is the gating condition
+  evaluation context, which may or may not have been mutated during this evaluation.
   """
-  @callback open?(GatingCondition.t()) :: Boolean.t()
+  @callback evaluate(GatingCondition.t(), ConditionContext.t()) ::
+              {Boolean.t(), ConditionContext.t()}
 
   @doc """
   Returns a human readable message as to why a resource access has been blocked.